requirements.txt
fastapi
uvicorn[standard]
pydantic
PyNaCl
prometheus_fastapi_instrumentator
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-asgi
opentelemetry-exporter-otlp
